+++ src/components.d.ts
@@ -59,9 +59,9 @@ export namespace Components {
     }
     interface MvxCopyButton {
         "class"?: string;
-        "copyIcon"?: IconDefinition;
+        "copyIcon"?: IconDefinition | string;
         "iconClass"?: string;
-        "successIcon"?: IconDefinition;
+        "successIcon"?: IconDefinition | string;
         "text": string;
     }
     interface MvxCustomToast {
@@ -74,7 +74,7 @@ export namespace Components {
     interface MvxExplorerLink {
         "class"?: string;
         "dataTestId"?: string;
-        "icon"?: IconDefinition;
+        "icon"?: IconDefinition | string;
         "iconClass"?: string;
         "link": string;
         "text"?: string;
@@ -1130,9 +1130,9 @@ declare namespace LocalJSX {
     }
     interface MvxCopyButton {
         "class"?: string;
-        "copyIcon"?: IconDefinition;
+        "copyIcon"?: IconDefinition | string;
         "iconClass"?: string;
-        "successIcon"?: IconDefinition;
+        "successIcon"?: IconDefinition | string;
         "text"?: string;
     }
     interface MvxCustomToast {
@@ -1146,7 +1146,7 @@ declare namespace LocalJSX {
     interface MvxExplorerLink {
         "class"?: string;
         "dataTestId"?: string;
-        "icon"?: IconDefinition;
+        "icon"?: IconDefinition | string;
         "iconClass"?: string;
         "link"?: string;
         "text"?: string;